AgentHubClient (AgentHub Client v1.0.0-alpha.2)

This is the main client module. It provides the top-level functions to interact with the AgentHub application, like querying the state of agents, fetching list of agents and sending messages to an agent.

Functions

count_by_voice_state(campaign_id)

Specs

count_by_voice_state(campaign_id :: String.t()) ::
  {:ok, %{optional(AgentHubClient.AgentState.voice_state()) => pos_integer()}}
  | {:error, reason :: String.t()}

Returns the number of agents on each voice state that has at least one agent, i.e., it won't include a state that has no agents in that state.

Examples

iex> AgentHubClient.count_by_voice_state("campaign1")
{:ok, %{idle: 1, talking: 1}}

iex> AgentHubClient.count_by_voice_state("inexistent")
{:ok, %{}}

get_all()

Specs

get_all() ::
  {:ok, [AgentHubClient.AgentState.agent_state()]}
  | {:error, reason :: String.t()}

Returns a list of all agents that are logged in, no matter the campaign.

get_all_in_campaign(campaign_id)

Specs

get_all_in_campaign(campaign_id :: String.t()) ::
  {:ok, [AgentHubClient.AgentState.agent_state()]}
  | {:error, reason :: String.t()}

Returns a list of all the agents that are logged in a given campaign.

get_by_user_id(user_id)

Specs

get_by_user_id(user_id :: String.t()) ::
  {:ok, AgentHubClient.AgentState.agent_state()}
  | {:error, reason :: String.t()}

Returns the state of the agent given its user ID.