AgentHubClient.AgentState (AgentHub Client v1.0.0-alpha.11)

Defines the agent state record to be used when dealing with mnesia and the agent state struct to be used in all the rest of the application.

Types

chat_state()

Specs

chat_state() :: :idle | :chatting | :wrap | :pause | :ending

An atom representing the agent's state related to chat operations.

t()

Specs

t() :: %AgentHubClient.AgentState{
  chat_state: chat_state(),
  chat_state_since: DateTime.t(),
  inserted_at: DateTime.t(),
  next_chat_state: chat_state(),
  next_voice_state: voice_state(),
  online_since: DateTime.t(),
  pause_code: String.t(),
  updated_at: DateTime.t(),
  user_id: String.t(),
  user_name: String.t(),
  voice_state: voice_state(),
  voice_state_since: DateTime.t()
}

A struct representing the state of an agent.

voice_state()

Specs

voice_state() ::
  :idle
  | :talking
  | :consulting
  | :holding
  | :transferring
  | :redial
  | :wrap
  | :pause
  | :ending
  | :private_callback
  | :personal_call
  | :manual_call

An atom representing the agent's state related to voice operations.

Functions

agent_state(args \\ [])
(macro)

agent_state(record, args)
(macro)

attributes()

The list of attributes. Used when creating Mnesia tables. See t/0.

from_record(arg1)

Specs

from_record(
  record ::
    {:agent_state, user_id :: term(), user_name :: term(),
     online_since :: term(), voice_state :: term(), voice_state_since :: term(),
     next_voice_state :: term(), chat_state :: term(),
     chat_state_since :: term(), next_chat_state :: term(),
     pause_code :: term(), inserted_at :: term(), updated_at :: term()}
) :: t()

Creates an AgentState struct from a record.

voice_states()

The list of voice states. Used to expose this enumeration to external modules.