AgentHubClient.Campaign (AgentHub Client v1.0.0-alpha.10)

Defines the campaign record to be used when dealing with mnesia and the campaign struct to be used in all the rest of the application.

Types

campaign_status()

Specs

campaign_status() :: :started | :stopped | :unknown

An atom representing the campaign's status.

t()

Specs

t() :: %AgentHubClient.Campaign{
  agents: [String.t()],
  id: String.t(),
  name: String.t(),
  started_at: DateTime.t(),
  status: campaign_status(),
  stopped_at: DateTime.t(),
  updated_at: DateTime.t()
}

A struct representing a campaign.

Functions

attributes()

campaign(args \\ [])
(macro)

campaign(record, args)
(macro)

from_map(map)

from_record(arg)