AgentHubClient.Cluster (AgentHub Client v0.0.30)

Provides supporting functions for dealing with node names, like identifying the application in which the client is running based on the node name.

Functions

agenthub?(node \\ Node.self())

Returns true iff the given node atom looks like an AgentHub node. If no node is given, assumes the current node.

Examples

iex> AgentHubClient.Cluster.agenthub? :"agenthub-xkpto123-blabla1@192.168.0.15"
true

iex> AgentHubClient.Cluster.agenthub? :"nonode@nohost"
false

iex> AgentHubClient.Cluster.agenthub? :"dialer@192.168.0.15"
false

connected_to_agenthub?()

Returns true iff the current node is connected to at least one AgentHub node

get_host(node)

Returns the host part of the given node atom (like the one from Node.self/0) as a string.

Examples

iex> AgentHubClient.Cluster.get_host :"agenthub1@127.0.0.1"
"127.0.0.1"

iex> AgentHubClient.Cluster.get_host :"nonode@nohost"
"nohost"

parse_node(node)

Parses the given node atom (like the one from Node.self/0) and returns its individual components as strings.

Examples

iex> AgentHubClient.Cluster.parse_node :"agenthub1@127.0.0.1"
["agenthub1", "127.0.0.1"]

iex> AgentHubClient.Cluster.parse_node :"nonode@nohost"
["nonode", "nohost"]

standalone?()