AgentHubClient.Store (AgentHub Client v0.0.31)

Provides an interface for interacting with the data store. In this version the data store is mnesia so we delegate all the functions to mnesia-specific modules.

Functions

change_voice_status(user_id, voice_status)

See AgentHubClient.Mnesia.Updates.change_voice_status/2.

count_by_voice_state(campaign_id, voice_state)

See AgentHubClient.Mnesia.Queries.count_by_voice_state/2.

delete(user_id)

See AgentHubClient.Mnesia.Delete.delete/1.

get_all()

See AgentHubClient.Mnesia.Queries.get_all/0.

get_all_in_campaign(campaign_id)

See AgentHubClient.Mnesia.Queries.get_all_in_campaign/1.

get_by_user_id(user_id)

See AgentHubClient.Mnesia.Queries.get_by_user_id/1.

get_by_voice_state(campaign_id, voice_state)

Specs

get_by_voice_state(String.t(), AgentHubClient.AgentState.voice_state()) :: [
  AgentHubClient.AgentState.agent_state()
]

See AgentHubClient.Mnesia.Queries.get_by_voice_state/2.

init(opts)

See AgentHubClient.Mnesia.Init.init/1.

insert(user_id, campaigns)

See AgentHubClient.Mnesia.Insert.insert/2.

set_offline(user_id)

See AgentHubClient.Mnesia.Updates.set_offline/1.

set_online(user_id)

See AgentHubClient.Mnesia.Updates.set_online/1.