AgentHubClient.Supervisor (AgentHub Client v1.0.0-alpha.12)

Supervises the internal processes used by the client itself (mnesia and libcluster).

Add it to your supervision tree (usually in application.ex). Example:

children = [
  {AgentHubClient.Supervisor,
   standalone: false,
   k8s_services: ["agenthub", "dialer"],
   replication: true}
]
Supervisor.start_link(children, strategy: :one_for_one)

All parameters are optional:

standalone: defaults to false, which means it will try to connect to remote nodes to form a cluster. Set it to true for running it in standalone mode, i.e., without connecting to any other node.

k8s_services: list of kubernetes services that are part of the cluster. It defaults to ["agenthub"], but it is ignored if running in standalone mode. Leave it blank if running multiple instances in a single node without kubernetes. In that case, different instances in the same host will see each other automatically. For production, add all the services that shall share data. Imagine we add only agenthub to this list and we have all the applications up and running. Then we restart the AgentHub. When it starts, it won't try to connect to the other applications, so it will start with a new empty table, while the dialer for instance might have a copy of the original table with data in it that we could use. If we add the dialer (and the other applications) here, then when the AgentHub starts, it will start connected to the dialer for instance and will join the cluster, copying the tables instead of creating new ones.

replication: set it to true to have a copy of the mnesia tables locally in this node, so reading data will be a local operation. If set to false, all access to the data will be remote as the data will be in the remote nodes only.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(init_arg)