AgentHubClient.Supervisor (AgentHub Client v1.0.0-alpha.2)

Supervises the internal processes used by the client itself.

Add it to your supervision tree (usually in application.ex). Example:

  children = [
    {AgentHubClient.Supervisor, replication: true}
  ]
  Supervisor.start_link(children, strategy: :one_for_one)

The only available option is replication. When enabled, a copy of the mnesia table that holds agents state is kept locally in this node, so reading data is a local operation. If disabled, all access to the data is remote as the data is in the AgentHub node.

Currently the internal processes that are supervised are mnesia and libcluster, where the libcluster process is only started if the node is started with a name, i.e, if it is started expecting it to be in a cluster and not standalone.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(init_arg)